# Use faster Rng in RandomIdGenerator (0%-6% performance improvement)