--- a/opentelemetry-sdk/Cargo.toml
+++ b/opentelemetry-sdk/Cargo.toml
@@ -21,7 +21,7 @@ futures-util = { version = "0.3.17", default-features = false, features = ["std"
 once_cell = "1.10"
 ordered-float = "4.0"
 percent-encoding = { version = "2.0", optional = true }
-rand = { version = "0.8", default-features = false, features = ["std", "std_rng"], optional = true }
+rand = { version = "0.8", default-features = false, features = ["std", "std_rng","small_rng"], optional = true }
 glob = {version = "0.3.1", optional =true}
 serde = { version = "1.0", features = ["derive", "rc"], optional = true }
 serde_json = { version = "1", optional = true }
@@ -68,6 +68,7 @@ harness = false
 [[bench]]
 name = "trace"
 harness = false
+required-features = ["testing"]
 
 [[bench]]
 name = "batch_span_processor"