@hdost 2023-06-09
😓 I need to re-run the baseline I realize my orignal baseline was on battery and I believe that may have shown the drastic improvements.