@djc 2023-06-09
Great to see some movement on this! Would be good to redo your measurements with a consistent environment. :smile:

Questions:

- Do we need a CSPRNG here, or is something insecure sufficient for our needs here?
- If we don't need a CSPRNG, we should probably just use `SmallRng`?
- If we do need a CSPRNG, how does rand_pcg compare to `ThreadRng`? IMO in the absence of strong motivation I would feel better sticking to rand's default algorithm selection.
- If rand_pcg provides a substantial advantage, can we use `Pcg64Mcg` on 64-bit architectures and `Pcg32` on 32-bit?