@hdost 2023-06-09
Still seeing improvements, but a bit more variable (2%-12%)
What is a bit curious from my perspective is the regression in the unrelated `EvictedHashMap` because it doesn't even use the `TraceIdGenerator`

I ran base lines a few times to normalize the temperature of my laptop and I did the same with the modification.
```
at 12:45:24 ➜  cargo bench -p opentelemetry_sdk --bench trace -- --baseline main
    Finished bench [optimized + debuginfo] target(s) in 0.13s
     Running benches/trace.rs (/home/h.dost/projects/github.com/hdost/opentelemetry-rust/target/release/deps/trace-eadd52583edeb932)
Gnuplot not found, using plotters backend
EvictedHashMap/insert 1 time:   [94.327 ns 95.895 ns 97.622 ns]
                        change: [+6.8380% +9.7777% +13.308%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
EvictedHashMap/insert 5 time:   [269.19 ns 271.21 ns 273.44 ns]
                        change: [-9.7396% -8.1467% -6.6378%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
EvictedHashMap/insert 10
                        time:   [562.49 ns 568.25 ns 574.97 ns]
                        change: [-11.395% -9.4004% -7.0972%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe
EvictedHashMap/insert 20
                        time:   [1.4855 µs 1.5105 µs 1.5351 µs]
                        change: [-3.6807% -1.3490% +1.1990%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe

start-end-span/always-sample
                        time:   [437.20 ns 439.62 ns 442.82 ns]
                        change: [-3.6715% -2.9462% -2.3529%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high severe
start-end-span/never-sample
                        time:   [129.74 ns 131.71 ns 133.63 ns]
                        change: [-12.133% -10.492% -8.8450%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

start-end-span-4-attrs/always-sample
                        time:   [1.1549 µs 1.1620 µs 1.1693 µs]
                        change: [-2.2722% -0.8020% +1.0240%] (p = 0.39 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
start-end-span-4-attrs/never-sample
                        time:   [162.44 ns 163.67 ns 165.07 ns]
                        change: [-17.960% -16.329% -14.688%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

start-end-span-8-attrs/always-sample
                        time:   [1.7150 µs 1.7223 µs 1.7299 µs]
                        change: [-3.2110% -1.9454% -0.5185%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
start-end-span-8-attrs/never-sample
                        time:   [209.97 ns 214.02 ns 217.95 ns]
                        change: [-10.995% -9.1955% -7.3323%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

start-end-span-all-attr-types/always-sample
                        time:   [1.4226 µs 1.4318 µs 1.4410 µs]
                        change: [+0.5991% +1.6015% +2.5197%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
start-end-span-all-attr-types/never-sample
                        time:   [174.47 ns 176.15 ns 178.10 ns]
                        change: [-14.156% -12.944% -11.760%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

start-end-span-all-attr-types-2x/always-sample
                        time:   [2.0565 µs 2.0719 µs 2.0903 µs]
                        change: [-3.9848% -2.7611% -1.3589%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
start-end-span-all-attr-types-2x/never-sample
                        time:   [223.16 ns 224.39 ns 225.64 ns]
                        change: [-13.450% -12.519% -11.496%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
```