@djc 2023-06-09
> > * If rand_pcg provides a substantial advantage, can we use `Pcg64Mcg` on 64-bit architectures and `Pcg32` on 32-bit?
> 
> I think we could just opt for the Pcg32.

If I'm reading the docs right, using `Pcg32` on 64-bit architectures would be bad for performance.