@djc 2023-06-09
> 



> From reading the documentation from the random team it does seem like SmallRng is not portable which I think we may wnat to avoid for better univerality.

What do you mean by portable, and universality? I don't see these terms in the documentation at https://docs.rs/rand/latest/rand/rngs/struct.SmallRng.html.