@hdost 2023-06-11
So i did a few tests looks like 2-7%. I'll post the full results when I'm back at my computer. 

For the entropy, i like the idea, but the one question I have is will this sharing of the first 64 bits result in a skew in of sampling.